+++ mayan/apps/document_signatures/urls.py
@@ -2,22 +2,40 @@ from __future__ import unicode_literals
 
 from django.conf.urls import patterns, url
 
+from .views import (
+    DocumentVersionSignatureDeleteView, DocumentVersionSignatureDetailView,
+    DocumentVersionSignatureListView
+)
+
 urlpatterns = patterns(
     'document_signatures.views',
     url(
-        r'^verify/(?P<document_pk>\d+)/$', 'document_verify',
-        name='document_verify'
+        r'^(?P<pk>\d+)/details/$',
+        DocumentVersionSignatureDetailView.as_view(),
+        name='document_version_signature_details'
     ),
     url(
-        r'^upload/signature/(?P<document_pk>\d+)/$',
-        'document_signature_upload', name='document_signature_upload'
+        r'^signature/(?P<pk>\d+)/download/$',
+        'document_signature_download',
+        name='document_version_signature_download'
     ),
     url(
-        r'^download/signature/(?P<document_pk>\d+)/$',
-        'document_signature_download', name='document_signature_download'
+        r'^document/version/(?P<pk>\d+)/signatures/list/$',
+        DocumentVersionSignatureListView.as_view(),
+        name='document_version_signature_list'
     ),
     url(
-        r'^document/(?P<document_pk>\d+)/signature/delete/$',
-        'document_signature_delete', name='document_signature_delete'
+        r'^documents/version/(?P<document_pk>\d+)/signature/verify/$',
+        'document_verify', name='document_version_signature_verify'
+    ),
+    url(
+        r'^documents/version/(?P<pk>\d+)/signature/upload/$',
+        'document_version_signature_upload',
+        name='document_version_signature_upload'
+    ),
+    url(
+        r'^signature/(?P<pk>\d+)/delete/$',
+        DocumentVersionSignatureDeleteView.as_view(),
+        name='document_version_signature_delete'
     ),
 )