+++ mayan/apps/document_signatures/views.py
@@ -14,19 +14,93 @@ from django.utils.html import escape
 from django.utils.translation import ugettext_lazy as _
 
 from acls.models import AccessControlList
+from common.generics import (
+    SingleObjectDeleteView, SingleObjectDetailView, SingleObjectListView
+)
 from django_gpg.literals import SIGNATURE_STATE_NONE, SIGNATURE_STATES
-from documents.models import Document
+from documents.models import Document, DocumentVersion
 from filetransfers.api import serve_file
 from permissions import Permission
 
-from .forms import DetachedSignatureForm
-from .models import DocumentVersionSignature
+from .forms import DetachedSignatureForm, DocumentVersionSignatureDetailForm
+from .models import DetachedSignature, SignatureBaseModel
 from .permissions import (
-    permission_document_verify, permission_signature_upload,
-    permission_signature_download, permission_signature_delete
+    permission_document_version_signature_view,
+    permission_document_version_signature_verify,
+    permission_document_version_signature_upload,
+    permission_document_version_signature_download,
+    permission_document_version_signature_delete
 )
 
 logger = logging.getLogger(__name__)
+
+
+class DocumentVersionSignatureDeleteView(SingleObjectDeleteView):
+    model = DetachedSignature
+
+    def get_extra_context(self):
+        return {
+            'document':  self.get_object().document_version.document,
+            'document_version': self.get_object().document_version,
+            'navigation_object_list': ('document', 'document_version', 'signature'),
+            'signature': self.get_object(),
+            'title': _('Delete detached signature: %s') % self.get_object()
+        }
+
+    def get_post_action_redirect(self):
+        return reverse(
+            'signatures:document_version_signature_list',
+            args=(self.get_object().document_version.pk,)
+        )
+
+
+class DocumentVersionSignatureDetailView(SingleObjectDetailView):
+    form_class = DocumentVersionSignatureDetailForm
+
+    def get_extra_context(self):
+        return {
+            'document': self.get_object().document_version.document,
+            'document_version': self.get_object().document_version,
+            'signature': self.get_object(),
+            'navigation_object_list': ('document', 'document_version', 'signature'),
+            'hide_object': True,
+            'title': _(
+                'Details for signature: %s'
+            ) % self.get_object(),
+        }
+
+    def get_queryset(self):
+        return SignatureBaseModel.objects.select_subclasses()
+
+
+class DocumentVersionSignatureListView(SingleObjectListView):
+    def get_document_version(self):
+        return get_object_or_404(DocumentVersion, pk=self.kwargs['pk'])
+
+    def get_extra_context(self):
+        return {
+            'document': self.get_document_version().document,
+            'document_version': self.get_document_version(),
+            'navigation_object_list': ('document', 'document_version'),
+            'hide_object': True,
+            'title': _(
+                'Signatures for document version: %s'
+            ) % self.get_document_version(),
+        }
+
+    def get_queryset(self):
+        queryset = self.get_document_version().signatures.all()
+
+        try:
+            Permission.check_permissions(
+                self.request.user, (permission_document_version_signature_view,)
+            )
+        except PermissionDenied:
+            return AccessControlList.objects.filter_by_access(
+                permission_document_version_signature_view, self.request.user, queryset
+            )
+        else:
+            return queryset
 
 
 def document_verify(request, document_pk):
@@ -84,19 +158,20 @@ def document_verify(request, document_pk):
     }, context_instance=RequestContext(request))
 
 
-def document_signature_upload(request, document_pk):
-    document = get_object_or_404(Document, pk=document_pk)
+
+def document_version_signature_upload(request, pk):
+    document_version = get_object_or_404(DocumentVersion, pk=pk)
 
     try:
         Permission.check_permissions(
-            request.user, (permission_signature_upload,)
+            request.user, (permission_document_version_signature_upload,)
         )
     except PermissionDenied:
         AccessControlList.objects.check_access(
-            permission_signature_upload, request.user, document
+            permission_document_version_signature_upload, request.user, document_version.document
         )
 
-    document.add_as_recent_document_for_user(request.user)
+    document_version.document.add_as_recent_document_for_user(request.user)
 
     post_action_redirect = None
     previous = request.POST.get('previous', request.GET.get('previous', request.META.get('HTTP_REFERER', reverse(settings.LOGIN_REDIRECT_URL))))
@@ -106,8 +181,9 @@ def document_signature_upload(request, document_pk):
         form = DetachedSignatureForm(request.POST, request.FILES)
         if form.is_valid():
             try:
-                DocumentVersionSignature.objects.add_detached_signature(
-                    document.latest_version, request.FILES['file']
+                DetachedSignature.objects.upload_signature(
+                    document_version=document_version,
+                    signature_file=request.FILES['file']
                 )
                 messages.success(
                     request, _('Detached signature uploaded successfully.')
@@ -122,9 +198,11 @@ def document_signature_upload(request, document_pk):
     return render_to_response('appearance/generic_form.html', {
         'form': form,
         'next': next,
-        'object': document,
+        'document': document_version.document,
+        'document_version': document_version,
+        'navigation_object_list': ('document', 'document_version'),
         'previous': previous,
-        'title': _('Upload detached signature for document: %s') % document,
+        'title': _('Upload detached signature for document version: %s') % document_version,
     }, context_instance=RequestContext(request))
 
 
@@ -133,11 +211,11 @@ def document_signature_download(request, document_pk):
 
     try:
         Permission.check_permissions(
-            request.user, (permission_signature_download,)
+            request.user, (permission_document_version_signature_download,)
         )
     except PermissionDenied:
         AccessControlList.objects.check_access(
-            permission_signature_download, request.user, document
+            permission_document_version_signature_download, request.user, document
         )
 
     try:
@@ -156,49 +234,3 @@ def document_signature_download(request, document_pk):
         return HttpResponseRedirect(request.META['HTTP_REFERER'])
 
     return HttpResponseRedirect(request.META['HTTP_REFERER'])
-
-
-def document_signature_delete(request, document_pk):
-    document = get_object_or_404(Document, pk=document_pk)
-
-    try:
-        Permission.check_permissions(
-            request.user, (permission_signature_delete,)
-        )
-    except PermissionDenied:
-        AccessControlList.objects.check_access(
-            permission_signature_delete, request.user, document
-        )
-
-    document.add_as_recent_document_for_user(request.user)
-
-    post_action_redirect = None
-    previous = request.POST.get('previous', request.GET.get('previous', request.META.get('HTTP_REFERER', reverse(settings.LOGIN_REDIRECT_URL))))
-    next = request.POST.get('next', request.GET.get('next', post_action_redirect if post_action_redirect else request.META.get('HTTP_REFERER', reverse(settings.LOGIN_REDIRECT_URL))))
-
-    if request.method == 'POST':
-        try:
-            DocumentVersionSignature.objects.clear_detached_signature(
-                document.latest_version
-            )
-            messages.success(
-                request, _('Detached signature deleted successfully.')
-            )
-            return HttpResponseRedirect(next)
-        except Exception as exception:
-            messages.error(
-                request, _(
-                    'Error while deleting the detached signature; %s'
-                ) % exception
-            )
-            return HttpResponseRedirect(previous)
-
-    return render_to_response('appearance/generic_confirm.html', {
-        'delete_view': True,
-        'next': next,
-        'object': document,
-        'previous': previous,
-        'title': _(
-            'Delete the detached signature from document: %s?'
-        ) % document,
-    }, context_instance=RequestContext(request))